requirements.txt
torch>=1.9.0
transformers>=4.30.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
tqdm>=4.62.0
jupyter>=1.0.0
accelerate>=0.20.0
pyarrow>=7.0.0
pyyaml>=6.0
matplotlib>=3.4.0
seaborn>=0.11.0
ipykernel>=6.0.0
notebook>=6.4.0 